refactor history after update
