+++ src/modules/history/actions.js
@@ -2,13 +2,23 @@ import API from '../../services/api';
 import { types } from './mutations';
 
 const actions = {
+  fetchAll: (store, { assetsIds, baseId, daysArr }) => {
+    store.commit(types.FETCH_PRICES_HISTORY_REQUEST, { baseId });
+    const promises = daysArr.map(days => actions.fetch(store, { assetsIds, baseId, days }));
+    return Promise.all(promises).then(() => {
+      store.commit(types.FETCH_PRICES_HISTORY_COMPLETE);
+    }).catch(err => {
+      console.log(err);
+      store.commit(types.FETCH_PRICES_HISTORY_ERROR);
+    });
+  },
   fetch: (store, { assetsIds, baseId, days }) => {
     const { commit, rootGetters } = store;
     const assets = rootGetters['assets/getAssets'];
     const baseAsset = assets[baseId];
 
     commit(types.FETCH_PRICES_HISTORY_REQUEST, { baseId });
-    Promise.all(assetsIds.map(async (assetId) => {
+    return Promise.all(assetsIds.map(async (assetId) => {
       const prices = await API.Assets.fetchPriceHistory(baseAsset, assets[assetId], days);
       if (!prices) throw new Error('error market history');
       return {
@@ -20,9 +30,8 @@ const actions = {
         result[obj.assetId] = obj.prices;
         return result;
       }, {});
-      commit(types.FETCH_PRICES_HISTORY_COMPLETE, { days, prices });
+      commit(types.FETCH_PRICES_HISTORY_UPDATE, { days, prices });
     }).catch((err) => {
-      commit(types.FETCH_PRICES_HISTORY_ERROR);
       console.log(err);
     });
   }
