+++ src/modules/history/defaultState.js
@@ -2,6 +2,7 @@ export const getDefaultState = () => {
   return {
     days: {},
     fetching: false,
+    initLoaded: false,
     error: false,
     baseId: ''
   };
