+++ src/modules/history/mutations.js
@@ -3,7 +3,8 @@ import Vue from 'vue';
 export const types = {
   FETCH_PRICES_HISTORY_REQUEST: 'FETCH_PRICES_HISTORY_REQUEST',
   FETCH_PRICES_HISTORY_COMPLETE: 'FETCH_PRICES_HISTORY_COMPLETE',
-  FETCH_PRICES_HISTORY_ERROR: 'FETCH_PRICES_HISTORY_ERROR'
+  FETCH_PRICES_HISTORY_ERROR: 'FETCH_PRICES_HISTORY_ERROR',
+  FETCH_PRICES_HISTORY_UPDATE: 'FETCH_PRICES_HISTORY_UPDATE'
 };
 
 export const mutations = {
@@ -11,9 +12,12 @@ export const mutations = {
     state.fetching = true;
     state.baseAssetId = baseId;
   },
-  [types.FETCH_PRICES_HISTORY_COMPLETE](state, { prices, days }) {
-    state.fetching = false;
+  [types.FETCH_PRICES_HISTORY_UPDATE](state, { prices, days }) {
     Vue.set(state.days, days, prices);
+  },
+  [types.FETCH_PRICES_HISTORY_COMPLETE](state) {
+    state.fetching = false;
+    state.initLoaded = true;
   },
   [types.FETCH_PRICES_HISTORY_ERROR](state) {
     state.fetching = false;
